aider: Replaced all occurrences of 'relation' with 'incidenceRel' in index.html to fix undefined variable error.
# Aider chat conversation:

USER: Uncaught ReferenceError: relation is not defined
    updateData file:///Users/akurz/alexhkurz-at-git/FCA/index.html:26
    <anonymous> file:///Users/akurz/alexhkurz-at-git/FCA/index.html:49
index.html:26:83
    updateData file:///Users/akurz/alexhkurz-at-git/FCA/index.html:26
    <anonymous> file:///Users/akurz/alexhkurz-at-git/FCA/index.html:49

ASSISTANT: The error message indicates that the variable `relation` is not defined in the `updateData` function in `index.html`. This is because the variable was renamed to `incidenceRel` in `main.js`, but the `updateData` function in `index.html` is still trying to access `relation`.

To fix this, we need to replace all occurrences of `relation` with `incidenceRel` in the `updateData` function in `index.html`.

Here is the *SEARCH/REPLACE* block:

index.html
```html
<<<<<<< SEARCH
    document.getElementById('attributesDisplay').innerText = "Attributes: " + relation.getAttributes().join(', ');
    document.getElementById('objectsDisplay').innerText = "Objects: " + relation.getObjects().join(', ');
=======
    document.getElementById('attributesDisplay').innerText = "Attributes: " + incidenceRel.getAttributes().join(', ');
    document.getElementById('objectsDisplay').innerText = "Objects: " + incidenceRel.getObjects().join(', ');
>>>>>>> REPLACE
```