aider: Modified down() function to accept weightedUpset as an argument and updated its call in index.html.
# Aider chat conversation:

USER: now i want to change the function down() so that it takes an argument.

ASSISTANT: Sure, I can help with that. To clarify, what should the argument to the `down()` function be?

USER: in our case it should be the weighted upset

ASSISTANT: Alright, I will modify the `down()` function in `main.js` to take `weightedUpset` as an argument.

Here is the *SEARCH/REPLACE* block:

main.js
```javascript
<<<<<<< SEARCH
function down() {
=======
function down(weightedUpset) {
>>>>>>> REPLACE
```

And we also need to update the call to `down()` in the `calculate()` function in `index.html` to pass `weightedUpset` as an argument.

index.html
```html
<<<<<<< SEARCH
let downUpResult = down(); // the downset of the upset
=======
let downUpResult = down(weightedUpset); // the downset of the upset
>>>>>>> REPLACE
```